[otbn] Factor KnownMem out of Model.py in RIG
No functional change, but this was getting a little unwieldy.
